requirements.txt
requests
PyJWT
python-dotenv
cryptography
